backend/requirements.txt
fastapi
uvicorn
korean_lunar_calendar
pydantic

ephem

python-dotenv

google-generativeai

gunicorn
